Revert "Rollup merge of #78934 - DeveloperC286:issue_60302_vec, r=m-ou-se"
This reverts commit 242a252687264e1115dbfa5f2b4890a35d224a05, reversing
changes made to 507bff92fadf1f25a830da5065a5a87113345163.
